Finishing es translations.
I've finished the translations to Spanish.

My translations are well suited for Colombian Spanish. They'll work well
for almost any Spanish, but I don't know how some words are translated
in other countries (e.g. Array is translated to es-co as arreglo.
However, it might be left as it is for other countries and the meaning
is well transmited).